@@ -1,4 +1,4 @@
-import { Fragment, useState } from 'react';
+import { useState } from 'react';
 import Header from './components/Layout/Header';
 import Meals from './components/Meals/Meals';
 import Cart from './components/Cart/Cart';
@@ -7,18 +7,14 @@ import CartProvider from './store/CartProvider';
 function App() {
   const [showCart, setShowCart] = useState(false);
 
-  const showCartHandler = () => {
-    setShowCart(true);
-  };
-
-  const hideCartHandler = () => {
-    setShowCart(false);
-  };
+  function showCartHandler() {
+    setShowCart((showCart) => !showCart);
+  }
 
   return (
     // wrapping the all components with the CartProvider component will give access to the cart-context data to all components
     <CartProvider>
-      {showCart && <Cart onClose={hideCartHandler} />}
+      {showCart && <Cart onShowCart={showCartHandler} />}
       <Header onShowCart={showCartHandler} />
       <main>
         <Meals />
