@@ -1,20 +1,19 @@
-import { Fragment } from 'react';
 import mealsImage from '../../assets/meals.jpg';
-import styles from './Header.module.css';
+import classes from './Header.module.css';
 import HeaderCartButton from './HeaderCartButton';
 
-const Header = props => {
+function Header({ onShowCart }) {
   return (
-    <Fragment>
-      <header className={styles.header}>
+    <>
+      <header className={classes.header}>
         <h1>React Meals</h1>
-        <HeaderCartButton onClose={props.onShowCart} />
+        <HeaderCartButton onShowCart={onShowCart} />
       </header>
-      <div className={styles['main-image']}>
-        <img src={mealsImage} alt="A table full of delicious food!" />
+      <div className={classes['main-image']}>
+        <img src={mealsImage} alt='A table full of delicious food!' />
       </div>
-    </Fragment>
+    </>
   );
-};
+}
 
 export default Header;
