@@ -1,52 +1,52 @@
 import Card from '../UI/Card';
-import styles from './AvailableMeals.module.css';
+import classes from './AvailableMeals.module.css';
 import ListItem from './MealItem/ListItem';
 
 const DUMMY_MEALS = [
-	{
-		id: 'm1',
-		name: 'Sushi',
-		description: 'Finest fish and veggies',
-		price: 22.99,
-	},
-	{
-		id: 'm2',
-		name: 'Schnitzel',
-		description: 'A german specialty!',
-		price: 16.5,
-	},
-	{
-		id: 'm3',
-		name: 'Barbecue Burger',
-		description: 'American, raw, meaty',
-		price: 12.99,
-	},
-	{
-		id: 'm4',
-		name: 'Green Bowl',
-		description: 'Healthy...and green...',
-		price: 18.99,
-	},
+  {
+    id: 'm1',
+    name: 'Sushi',
+    description: 'Finest fish and veggies',
+    price: 22.99,
+  },
+  {
+    id: 'm2',
+    name: 'Schnitzel',
+    description: 'A german specialty!',
+    price: 16.5,
+  },
+  {
+    id: 'm3',
+    name: 'Barbecue Burger',
+    description: 'American, raw, meaty',
+    price: 12.99,
+  },
+  {
+    id: 'm4',
+    name: 'Green Bowl',
+    description: 'Healthy...and green...',
+    price: 18.99,
+  },
 ];
 
-const AvailableMeals = (props) => {
-	const dummyMeals = DUMMY_MEALS.map((meal) => (
-		<ListItem
-			id={meal.id}
-			key={meal.id}
-			name={meal.name}
-			description={meal.description}
-			price={meal.price}
-		/>
-	));
+function AvailableMeals() {
+  const mealsList = DUMMY_MEALS.map((meal) => (
+    <ListItem
+      id={meal.id}
+      key={meal.id}
+      name={meal.name}
+      description={meal.description}
+      price={meal.price}
+    />
+  ));
 
-	return (
-		<section className={styles.meals}>
-			<Card>
-				<ul>{dummyMeals}</ul>
-			</Card>
-		</section>
-	);
-};
+  return (
+    <section className={classes.meals}>
+      <Card>
+        <ul>{mealsList}</ul>
+      </Card>
+    </section>
+  );
+}
 
 export default AvailableMeals;
