@@ -1,14 +1,13 @@
-import { Fragment } from 'react';
 import MealsSummary from './MealsSummary';
 import AvailableMeals from './AvailableMeals';
 
-const Meals = () => {
+function Meals() {
   return (
-    <Fragment>
+    <>
       <MealsSummary />
       <AvailableMeals />
-    </Fragment>
+    </>
   );
-};
+}
 
 export default Meals;
