@@ -1,8 +1,8 @@
-import styles from './MealsSummary.module.css';
+import classes from './MealsSummary.module.css';
 
-const MealsSummary = () => {
+function MealsSummary() {
   return (
-    <section className={styles.summary}>
+    <section className={classes.summary}>
       <h2>Delicious Food, Delivered To You</h2>
       <p>
         Choose your favorite meal from our broad selection of available meals
@@ -14,6 +14,6 @@ const MealsSummary = () => {
       </p>
     </section>
   );
-};
+}
 
 export default MealsSummary;
